apps/api/requirements.txt
# Python dependencies for RestoNext MX API

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# AI & Data Science
pandas==2.1.4
prophet==1.1.5

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
